remove unused REST endpoints
