+++ packages/backend/src/routes/index.ts
@@ -1,6 +1,4 @@
-export * from "./root";
 export * from "./courses";
 export * from "./assignments";
 export * from "./lectures";
-export * from "./timetable";
 export * from "./gql";
